fix: hud auto scrolling
Formerly the HUD window would not auto scroll to the latest text line
(in case the terminal buffer was bigger than the window) after it's
initial appearance. Instead the solution was using a predefined delay
which did not always worked correctly. Now new updates to the REPL
terminal are automatically scrolled into view for as long as the HUD is
visible on screen.